requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
asttokens=2.2.1=pypi_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.2=pypi_0
blas=1.0=mkl
ca-certificates=2023.05.30=h06a4308_0
certifi=2023.5.7=pypi_0
charset-normalizer=3.2.0=pypi_0
click=8.1.5=pypi_0
comm=0.1.3=pypi_0
contourpy=1.1.0=pypi_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.11.0=pypi_0
datasets=2.13.1=pypi_0
debugpy=1.6.7=pypi_0
decorator=5.1.1=pypi_0
dill=0.3.6=pypi_0
editdistpy=0.1.3=pypi_0
et-xmlfile=1.1.0=pypi_0
executing=1.2.0=pypi_0
filelock=3.12.2=pypi_0
fonttools=4.41.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=pypi_0
fsspec=2023.6.0=pypi_0
gdown=4.7.1=pypi_0
gensim=4.3.1=pypi_0
giflib=5.2.1=h5eee18b_3
huggingface-hub=0.16.4=pypi_0
idna=3.4=pypi_0
ijson=3.2.2=pypi_0
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.0.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46305
ipykernel=6.24.0=pypi_0
ipython=8.12.2=pypi_0
ipywidgets=8.0.7=pypi_0
jedi=0.18.2=pypi_0
joblib=1.3.1=pypi_0
jpeg=9e=h5eee18b_1
jupyter-client=8.3.0=pypi_0
jupyter-core=5.3.1=pypi_0
jupyterlab-widgets=3.0.8=pypi_0
kiwisolver=1.4.4=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.0=h6a678d5_2
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lz4-c=1.9.4=h6a678d5_0
matplotlib=3.7.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.6=py38h417a72b_1
mkl_random=1.2.2=py38h417a72b_1
multidict=6.0.4=pypi_0
multiprocess=0.70.14=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pypi_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
nlpaug=1.1.11=pypi_0
nltk=3.8.1=pypi_0
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
openpyxl=3.1.2=pypi_0
openssl=3.0.9=h7f8727e_0
packaging=23.1=pypi_0
pandas=2.0.3=pypi_0
parso=0.8.3=pypi_0
patsy=0.5.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=py38h6a678d5_0
pip=23.1.2=py38h06a4308_0
platformdirs=3.8.1=pypi_0
pprintpp=0.4.0=pypi_0
prompt-toolkit=3.0.39=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=12.0.1=pypi_0
pygments=2.15.1=pypi_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=pypi_0
python=3.8.17=h955ad1f_0
python-dateutil=2.8.2=pypi_0
pytorch=1.6.0=py3.8_cuda10.1.243_cudnn7.6.3_0
pytz=2023.3=pypi_0
pyyaml=6.0=pypi_0
pyzmq=25.1.0=pypi_0
readline=8.2=h5eee18b_0
regex=2023.6.3=pypi_0
requests=2.31.0=pypi_0
sacremoses=0.0.53=pypi_0
safetensors=0.3.1=pypi_0
scikit-learn=1.3.0=pypi_0
scipy=1.10.1=pypi_0
sentencepiece=0.1.99=pypi_0
setuptools=67.8.0=py38h06a4308_0
six=1.16.0=pypi_0
smart-open=6.3.0=pypi_0
soupsieve=2.4.1=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.2=pypi_0
statsmodels=0.14.0=pypi_0
symspellpy=6.7.7=pypi_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
torchvision=0.7.0=py38_cu101
tornado=6.3.2=pypi_0
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pypi_0
transformers=4.30.2=pypi_0
typing-extensions=4.7.1=pypi_0
tzdata=2023.3=pypi_0
urllib3=2.0.3=pypi_0
wcwidth=0.2.6=pypi_0
wheel=0.38.4=py38h06a4308_0
widgetsnbextension=4.0.8=pypi_0
wordninja=2.0.0=pypi_0
xxhash=3.2.0=pypi_0
xz=5.4.2=h5eee18b_0
yarl=1.9.2=pypi_0
zipp=3.16.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
